backend/requirements/requirements.dev.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.1.0
cryptography==44.0.1
defusedxml==0.7.1
Django==3.2
django-filter==2.4.0
django-link-url-shortener==1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.3.1
flake8==7.1.2
gunicorn==20.1.0
idna==3.10
mccabe==0.7.0
oauthlib==3.2.2
pillow==11.1.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==1.26.15
